+++ src/Namotion.Interceptor.Generator.Tests/InterfaceDefaultPropertyBehaviorTests.cs
@@ -1,10 +1,28 @@
-using Namotion.Interceptor.Generator.Tests.Models;
+using Namotion.Interceptor.Attributes;
 using Namotion.Interceptor.Registry;
 
 namespace Namotion.Interceptor.Generator.Tests;
 
-public class InterfaceDefaultPropertyBehaviorTests
+public partial class InterfaceDefaultPropertyBehaviorTests
 {
+    #region InterfaceDefaultProperty_RegisteredAndWorks
+
+    public interface ITemperatureSensorInterface
+    {
+        double TemperatureCelsius { get; set; }
+
+        [Derived]
+        double TemperatureFahrenheit => TemperatureCelsius * 9 / 5 + 32;
+
+        bool IsFreezing => TemperatureCelsius <= 0;
+    }
+
+    [InterceptorSubject]
+    public partial class SensorWithInterfaceDefaults : ITemperatureSensorInterface
+    {
+        public partial double TemperatureCelsius { get; set; }
+    }
+
     [Fact]
     public void InterfaceDefaultProperty_RegisteredAndWorks()
     {
@@ -34,6 +52,26 @@ public class InterfaceDefaultPropertyBehaviorTests
         Assert.Null(defaultProp.SetValue);
     }
 
+    #endregion
+
+    #region NestedInterface_RegisteredAndWorks
+
+    public partial class OuterClass
+    {
+        public interface INestedSensor
+        {
+            double Value { get; set; }
+
+            string Status => Value > 0 ? "Active" : "Inactive";
+        }
+    }
+
+    [InterceptorSubject]
+    public partial class SensorWithNestedInterface : OuterClass.INestedSensor
+    {
+        public partial double Value { get; set; }
+    }
+
     [Fact]
     public void NestedInterface_RegisteredAndWorks()
     {
@@ -54,6 +92,23 @@ public class InterfaceDefaultPropertyBehaviorTests
         var statusProp = SubjectPropertyMetadataCache.Get<SensorWithNestedInterface>()["Status"];
         Assert.False(statusProp.IsIntercepted);
         Assert.Equal("Active", statusProp.GetValue?.Invoke(sensor));
+    }
+
+    #endregion
+
+    #region WritableInterfaceDefault_RegisteredAndWorks
+
+    public interface IWritableDefaultInterface
+    {
+        double Temperature { get; set; }
+
+        string Label { get => $"Temp: {Temperature}"; set { } }
+    }
+
+    [InterceptorSubject]
+    public partial class SensorWithWritableDefault : IWritableDefaultInterface
+    {
+        public partial double Temperature { get; set; }
     }
 
     [Fact]
@@ -79,6 +134,23 @@ public class InterfaceDefaultPropertyBehaviorTests
         Assert.Null(labelProp.SetValue);
     }
 
+    #endregion
+
+    #region InitOnlyInterface_RegisteredAndWorks
+
+    public interface IInitOnlyInterface
+    {
+        string Id { get; init; }
+
+        string DisplayId => $"ID: {Id}";
+    }
+
+    [InterceptorSubject]
+    public partial class SensorWithInitOnly : IInitOnlyInterface
+    {
+        public partial string Id { get; init; }
+    }
+
     [Fact]
     public void InitOnlyInterface_RegisteredAndWorks()
     {
@@ -100,4 +172,6 @@ public class InterfaceDefaultPropertyBehaviorTests
         Assert.False(displayIdProp.IsIntercepted);
         Assert.Equal("ID: ABC123", displayIdProp.GetValue?.Invoke(sensor));
     }
+
+    #endregion
 }
