--- src/Namotion.Interceptor.Generator.Tests/Models/IInitOnlyInterface.cs
@@ -1,8 +1,0 @@
-namespace Namotion.Interceptor.Generator.Tests.Models;
-
-public interface IInitOnlyInterface
-{
-    string Id { get; init; }
-
-    string DisplayId => $"ID: {Id}";
-}
