--- src/Namotion.Interceptor.Generator.Tests/Models/IWritableDefaultInterface.cs
@@ -1,8 +1,0 @@
-namespace Namotion.Interceptor.Generator.Tests.Models;
-
-public interface IWritableDefaultInterface
-{
-    double Temperature { get; set; }
-
-    string Label { get => $"Temp: {Temperature}"; set { } }
-}
