--- src/Namotion.Interceptor.Generator.Tests/Models/SensorWithInitOnly.cs
@@ -1,9 +1,0 @@
-using Namotion.Interceptor.Attributes;
-
-namespace Namotion.Interceptor.Generator.Tests.Models;
-
-[InterceptorSubject]
-public partial class SensorWithInitOnly : IInitOnlyInterface
-{
-    public partial string Id { get; init; }
-}
