--- src/Namotion.Interceptor.Generator.Tests/Models/SensorWithInterfaceDefaults.cs
@@ -1,9 +1,0 @@
-using Namotion.Interceptor.Attributes;
-
-namespace Namotion.Interceptor.Generator.Tests.Models;
-
-[InterceptorSubject]
-public partial class SensorWithInterfaceDefaults : ITemperatureSensorInterface
-{
-    public partial double TemperatureCelsius { get; set; }
-}
