--- src/Namotion.Interceptor.Generator.Tests/Models/SensorWithNestedInterface.cs
@@ -1,9 +1,0 @@
-using Namotion.Interceptor.Attributes;
-
-namespace Namotion.Interceptor.Generator.Tests.Models;
-
-[InterceptorSubject]
-public partial class SensorWithNestedInterface : OuterClass.INestedSensor
-{
-    public partial double Value { get; set; }
-}
