--- src/Namotion.Interceptor.Generator.Tests/Models/SensorWithWritableDefault.cs
@@ -1,9 +1,0 @@
-using Namotion.Interceptor.Attributes;
-
-namespace Namotion.Interceptor.Generator.Tests.Models;
-
-[InterceptorSubject]
-public partial class SensorWithWritableDefault : IWritableDefaultInterface
-{
-    public partial double Temperature { get; set; }
-}
